Feat: Reserva

+++ vslviagens-api/src/main/java/com/vslviagens/vslviagensapi/dto/DetalhesVoosDTO.java
@@ -2,5 +2,5 @@ package com.vslviagens.vslviagensapi.dto;
 
 import com.vslviagens.vslviagensapi.model.ClasseVooRole;
 
-public record DetalhesVoosDTO(Long vooId, ClasseVooRole classe, double preco, double passageiros) {
+public record DetalhesVoosDTO(Long vooId, ClasseVooRole classe, double preco, int passageiros) {
 }
